pyproject.toml
[tool.poetry]
name = "ferrea-libraries"
version = "0.1.0"
description = "\"Microservice for Ferrea framework.\""
authors = ["Eugenio Grimoldi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
attrs = "^22.2.0"
ferrea = {git = "https://github.com/Grimoldi/pyferrea-chassis.git"}
uvicorn = "^0.27.1"
fastapi-utils = "^0.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
